# api/openapi/osai-v1.yaml
openapi: 3.0.3
info:
  title: Osai Customer API
  version: 1.0.0
  description: Phase 5 customer-facing API surface for quote, trade, settlement, balances, and transactions.
servers:
  - url: http://localhost:8082
paths:
...
  /v1/quotes:
    post:
      summary: Create a quote
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/RequestID'
        - $ref: '#/components/parameters/Authorization'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateQuoteRequest'
      responses:
        '200':
          description: Quote created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QuoteResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
  /v1/quotes/{quote_id}:
    get:
      summary: Get quote by ID
      parameters:
        - $ref: '#/components/parameters/Authorization'
        - $ref: '#/components/parameters/RequestID'
        - $ref: '#/components/parameters/QuoteID'
      responses:
        '200':
          description: Quote found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QuoteResponse'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: Authorization
      description: 'Format: ApiKey <public_id>:<secret>'
  parameters:
    Authorization:
      name: Authorization
      in: header
      required: true
      schema:
        type: string
    RequestID:
      name: X-Request-Id
      in: header
      required: false
      schema:
        type: string
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema:
        type: string
...
    QuoteID:
      name: quote_id
      in: path
      required: true
      schema:
        type: string
...
  schemas:
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, request_id]
          properties:
            code:
              type: string
            message:
              type: string
            request_id:
              type: string
...
    QuoteResponse:
      type: object
      required: [quote_id, status, base_amount_minor, base_currency, quote_currency, amount_out_minor, rate_minor, fee_minor, expires_at]
      properties:
        quote_id:
          type: string
        status:
          type: string
        base_amount_minor:
          type: integer
        base_currency:
          type: string
        quote_currency:
          type: string
        destination_rail:
          type: string
        amount_out_minor:
          type: integer
        rate_minor:
          type: integer
        fee_minor:
          type: integer
        expires_at:
          type: string
          format: date-time
        request_id:
          type: string
    CreateQuoteRequest:
      type: object
      required: [base_amount_minor, base_currency, quote_currency]
      properties:
        base_amount_minor:
          type: integer
        base_currency:
          type: string
        quote_currency:
          type: string
        destination_rail:
          type: string
        urgency:
          type: string
...
  responses:
    ValidationError:
      description: Validation failure
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
security:
  - ApiKeyAuth: []